Remove IsSeries property and Series option
It's not doing anything. I'd marked it with [YamlIgnore] for some reason, so it wasn't even updated in the header and written back. I'm not sure what I'd intended to use it for, and I can't think what it might be just from looking at the code and the blog posts which are in a series.

This also means I should remove the --series option, as it is also not doing anything.