Discovered a neater way of suppling the global option
Rather than passing the global option instance around as a ctor, I saw a trick where you define it in a static class. It seems a much neater way of passing it around sub-commands.

This also fixes #37 by introducing a reporter object for the InfoCommandHandler which runs the rendering code.