Use mocking for tests
In order to test the publish command handler, rather than check the console output, instead we can mock the ILogger and assert the right methods are called on it with the right parameters.

Add NSubsitute. This leads to quickly finding out that mocking ILogger isn't all that straightforward. Hence the addition of MockLogger, which I found on a github issue on the NSubstitute repo